+++ src/components/Container/index.tsx
@@ -1,0 +1,10 @@
+// import React from 'react'
+import styled from 'styled-components'
+
+const LayoutCenter = styled.div`
+  width: 100%;
+  max-width: 600px;
+  margin: 0 auto;
+`
+
+export default LayoutCenter
