+++ src/style.css
@@ -21,7 +21,6 @@ body {
 
 #root {
   height: 100%;
-  max-width: 600px;
   margin: auto;
 }
